phase 7 level 5: zero PHPStan errors without baseline
Function signature fixes and argument type corrections:

- conf_update_param @param string $value -> mixed (accepts array/int/bool/float)
- pwg_db_get_recent_period_expression: string -> int|string period param
- pwg_db_get_flood_period_expression: string -> int|string
- format_date: accept DateTime|int|string|null (handle DateTime directly)
- time_since: accept int|string|null (was ?string)
- verify_ephemeral_key: array -> string (function body uses explode on it)
- get_image_ids_for_tags $order_by: string -> string|null
- validate_mail_address $user_id: int -> ?int
- page_not_found: string -> ?string
- add_permission_on_category $user_ids: accept int|string|array
- delete_element_derivatives $type: accept string|int|'all'
- need_resize: string -> int|string for max_width/max_height
- get_subcat_ids: int[] -> int[]|string[]
- set_tags/add_tags: int[] -> int[]|string[]
- update_category: accept string[] ids
- find_tags: int[] -> int[]|string[]
- userprefs_update_param $value: string -> mixed
- get_themeconf: docblock @param array -> string

Type cast and argument fixes:
- Cast int to string for base_convert(), uniqid() calls
- Cast float to int for chr(), crop_h(), crop_v(), sharpen(), compose()
- Cast string to int for log_user(), calculate_auto_login_key()
- Cast to bool for curl_setopt() boolean options
- Cast to int for checkdate() args
- Cast to float for round() args
- Fix preg_replace_callback closures to return string (not int)
- Fix usort on Config::formatExtensions() to use variable reference
- Fix ws_init.inc.php: remove invalid int 4th arg to add_event_handler
- Fix mysqli port: cast string to int
- Fix setcookie: false -> ''
- Fix i.php: cast compose x/y params to int
- Fix install.php: cast time() to string for uniqid()
- Fix comments.php: pass false instead of '' to create_navigation_bar
- Fix PwgTemplateAdapter::derivative: create SrcImage from array
- Fix SizingParams::compute: add string|null to $coi type
- Fix fraction_to_char: cast chr arg to int
- Fix Logger::formatMessage: @param string -> int for $level
- Fix ScriptLoader::add: string -> string|null for path
- Fix FileCombiner: false -> 0 for Combinable version
- Fix add_event_handler: accept callable|string (not just callable())
- phpstan-nobaseline.neon now at level 5; all checks passing

+++ admin/include/functions_metadata.php
@@ -35,7 +35,7 @@ function get_sync_iptc_data($file): array
                 $month = $matches[2];
                 $day = $matches[3];
 
-                if (!checkdate($month, $day, $year)) {
+                if (!checkdate((int)$month, (int)$day, (int)$year)) {
                     // we suppose the year is correct
                     $month = 1;
                     $day = 1;
@@ -173,13 +173,13 @@ function get_sync_metadata($infos)
                 if (isset($width) and $width != '') {
                     $infos['width'] = (int) $width;
                 } elseif ($vb !== '') {
-                    $infos['width'] = round(explode(' ', $vb)[2]);
+                    $infos['width'] = round((float)explode(' ', $vb)[2]);
                 }
 
                 if (isset($height) and $height != '') {
                     $infos['height'] = (int) $height;
                 } elseif ($vb !== '') {
-                    $infos['height'] = round(explode(' ', $vb)[3]);
+                    $infos['height'] = round((float)explode(' ', $vb)[3]);
                 }
             }
             if ($image_size = @getimagesize($file)) {
